refactor: Rename helpers to mocks and move
