+++ src/app/core/services/mocks.ts
@@ -1,0 +1,39 @@
+import { BehaviorSubject } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { PackageList, Package, PackageVersion } from '../models/package.model';
+
+export class MockHttpClient {
+  get() {}
+  post() {}
+  delete() {}
+  put() {}
+}
+
+export class MockThemeService {
+  private isDarkModeSource = new BehaviorSubject<boolean>(false);
+  isDarkMode$ = this.isDarkModeSource.asObservable();
+  themeClass$ = this.isDarkModeSource
+    .asObservable()
+    .pipe(map(darkMode => (darkMode ? 'app-dark-theme' : 'app-light-theme')));
+  toggleDarkMode() {
+    this.isDarkModeSource.next(!this.isDarkModeSource.value);
+  }
+}
+
+export class MockPackageService {
+  installedPackages$ = new BehaviorSubject<PackageList>([]);
+
+  installPackage(pkg: Package, version: PackageVersion) {
+    this.installedPackages$.next([...this.installedPackages$.value, pkg]);
+  }
+
+  public uninstallPackage(pkg: Package) {
+    this.installedPackages$.next(
+      this.installedPackages$.value.filter(
+        installed => installed.uuid4 !== pkg.uuid4
+      )
+    );
+  }
+
+  public updatePackage(pkg: Package, version: PackageVersion) {}
+}
