+++ src/app/packages/package-table/package-table.component.ts
@@ -36,7 +36,6 @@ export class PackageTableComponent implements AfterViewInit, OnDestroy {
 
   constructor(
     public thunderstore: ThunderstoreService,
-    public service: PackageService
   ) {}
 
   ngAfterViewInit() {
