+++ src/app/packages/packages-page/packages-page.component.spec.ts
@@ -3,6 +3,8 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { PackagesPageComponent } from './packages-page.component';
 import { Component, Input } from '@angular/core';
 import { Package } from '../../core/models/package.model';
+import { PackageService } from '../../core/services/package.service';
+import { MockPackageService } from '../../core/services/mocks';
 
 @Component({
   selector: 'app-package-table',
@@ -20,7 +22,8 @@ describe('PackagesPageComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [PackagesPageComponent, MockPackageTableComponent]
+      declarations: [PackagesPageComponent, MockPackageTableComponent],
+      providers: [{ provide: PackageService, useClass: MockPackageService }]
     }).compileComponents();
   }));
