--- src/app/shared/helpers.ts
@@ -1,20 +1,0 @@
-import { BehaviorSubject } from 'rxjs';
-import { map } from 'rxjs/operators';
-
-export class MockHttpClient {
-  get() {}
-  post() {}
-  delete() {}
-  put() {}
-}
-
-export class MockThemeService {
-  private isDarkModeSource = new BehaviorSubject<boolean>(false);
-  isDarkMode$ = this.isDarkModeSource.asObservable();
-  themeClass$ = this.isDarkModeSource
-    .asObservable()
-    .pipe(map(darkMode => (darkMode ? 'app-dark-theme' : 'app-light-theme')));
-  toggleDarkMode() {
-    this.isDarkModeSource.next(!this.isDarkModeSource.value);
-  }
-}
